Sync plperl with PostgreSQL REL9_1_STABLE.
We had mostly backported plperl from 9.1, but not up to the tip of
the stable branch, and there were some cosmetic and other differences
that were not really necessary. It is simpler if the GPDB version is
as identical as possible to a particular upstream version, not a mismatch
of different versions.

Copy over everything from upstream REL9_1_STABLE, and tweak things to make
it compile and pass the regression tests.

I started hacking this because the regression tests were not passing. Now
they pass.